Typo fixed in: Update README.md
Hi,

I suggest two grammatical changes for this text:

First one: "it's" vs. "its":

"Each peer/node on the network is identified via it's ENR..."
→ "Each peer/node on the network is identified via its ENR..."
Explanation: "It's" is a contraction of "it is," whereas "its" is the possessive form.

Second one: "PING/PONG's":

"through regular PING/PONG's with discovered nodes."
→ "through regular PING/PONGs with discovered nodes."
Explanation: "PING/PONGs" should not have an apostrophe since it refers to a plural form, not possession.

Thanks and a happy new year.